Reset to default session configs to fix Oauth CSRF issue
We aren't actually using the domain: :all session option
for the People app (in theory we could try to set sessions
across all the *.turing.io domains but we don't really need
this for People).

Since this seems to be causing the session CSRF issue for omniauth,
let's just revert to normal settings.